hashtrade_ai/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
python-dotenv
openai
google-generativeai
requests
yfinance
email-validator
